move test functions to tests/
